Many small changes

+++ src/main/kotlin/spixie/renderer/AparapiKernelInterface.java
@@ -5,5 +5,5 @@ import com.aparapi.opencl.OpenCL;
 
 @OpenCL.Resource("kernel.cl")
 public interface AparapiKernelInterface extends OpenCL<AparapiKernelInterface> {
-    public AparapiKernelInterface renderParticles(Range _range, @GlobalReadOnly("particles") float[] particles, @Arg("width") int width, @Arg("height") int height, @Arg("realWidth") int realWidth, @Arg("particlesCount") int particlesCount, @GlobalWriteOnly("outImage") int[] outImage);
+    public AparapiKernelInterface renderParticles(Range _range, @GlobalReadOnly("particles") float[] particles, @Arg("width") int width, @Arg("height") int height, @Arg("realWidth") int realWidth, @Arg("particlesCount") int particlesCount, @GlobalWriteOnly("outImage") float[] outImage);
 }
